backend/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.7
beautifulsoup4==4.12.2
selenium==4.15.2
undetected-chrome==3.5.4
fake-useragent==1.4.0
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
scikit-learn==1.3.0
joblib==1.3.2
jinja2==3.1.2
functions-framework==3.4.0
google-cloud-secret-manager==2.16.4
google-cloud-sql-connector==1.4.3